requirements.txt
fastapi[all]
openai
google-cloud-texttospeech